refactor: reset password and signup
